# feat(intl-messageformat-parser): Rewrite grammar