--- a/package.json
+++ b/package.json
@@ -73,6 +73,7 @@
     "test262-harness": "^6.3.1",
     "ts-jest": "^24.0.2",
     "ts-node": "^8.1.0",
+    "ts-pegjs": "^0.2.5",
     "tslib": "^1.10.0",
     "typescript": "^3.4.5"
   },