--- a/packages/babel-plugin-react-intl/src/print-icu-message.ts
+++ b/packages/babel-plugin-react-intl/src/print-icu-message.ts
@@ -4,13 +4,23 @@
  * See the accompanying LICENSE file for terms.
  */
 
-import parser, {
-  MessageFormatPattern,
-  ElementFormat,
-  MessageTextElement,
-  SimpleFormat,
+import {
+  parse,
+  MessageFormatElement,
+  isLiteralElement,
+  LiteralElement,
   ArgumentElement,
-  PluralFormat
+  DateElement,
+  TimeElement,
+  NumberElement,
+  SelectElement,
+  PluralElement,
+  isSelectElement,
+  isArgumentElement,
+  isDateElement,
+  isTimeElement,
+  isNumberElement,
+  isPluralElement
 } from 'intl-messageformat-parser';
 
 const ESCAPED_CHARS: Record<string, string> = {
@@ -23,80 +33,68 @@ const ESCAPED_CHARS: Record<string, string> = {
 const ESAPE_CHARS_REGEXP = /\\#|[{}\\]/g;
 
 export default function printICUMessage(message: string) {
-  let ast = parser.parse(message);
+  let ast = parse(message);
   return _printICUMessage(ast);
 }
 
-function isPluralFormat(format: ElementFormat): format is PluralFormat {
-  return format.type === 'pluralFormat';
-}
-
-function isSimpleFormat(format: ElementFormat): format is SimpleFormat {
-  return (
-    format.type === 'numberFormat' ||
-    format.type === 'dateFormat' ||
-    format.type === 'timeFormat'
-  );
-}
-
-function _printICUMessage(ast: MessageFormatPattern): string {
-  let printedNodes = ast.elements.map(node => {
-    if (node.type === 'messageTextElement') {
-      return printMessageTextASTNode(node);
+function _printICUMessage(ast: MessageFormatElement[]): string {
+  let printedNodes = ast.map(el => {
+    if (isLiteralElement(el)) {
+      return printLiteralElement(el);
     }
 
-    if (!node.format) {
-      return `{${node.id}}`;
+    if (isArgumentElement(el)) {
+      return printArgumentElement(el);
+    }
+    if (isDateElement(el) || isTimeElement(el) || isNumberElement(el)) {
+      return printSimpleFormatElement(el);
     }
 
-    switch (getArgumentType(node.format)) {
-      case 'number':
-      case 'date':
-      case 'time':
-        return printSimpleFormatASTNode(node);
+    if (isPluralElement(el)) {
+      return printPluralElement(el);
+    }
 
-      case 'plural':
-      case 'selectordinal':
-      case 'select':
-        return printOptionalFormatASTNode(node);
+    if (isSelectElement(el)) {
+      return printSelectElement(el);
     }
   });
 
   return printedNodes.join('');
 }
 
-function getArgumentType(format: ElementFormat): string {
-  const { type } = format;
-
-  // Special-case ordinal plurals to use `selectordinal` instead of `plural`.
-  if (isPluralFormat(format) && format.ordinal) {
-    return 'selectordinal';
-  }
-
-  return type.replace(/Format$/, '').toLowerCase();
-}
-
-function printMessageTextASTNode({ value }: MessageTextElement) {
+function printLiteralElement({ value }: LiteralElement) {
   return value.replace(ESAPE_CHARS_REGEXP, char => ESCAPED_CHARS[char]);
 }
 
-function printSimpleFormatASTNode({ id, format }: ArgumentElement) {
-  let argumentType = getArgumentType(format);
-  let style = isSimpleFormat(format) && format.style ? `, ${format.style}` : '';
-
-  return `{${id}, ${argumentType}${style}}`;
+function printArgumentElement({ value }: ArgumentElement) {
+  return `{${value}}`;
 }
 
-function printOptionalFormatASTNode({ id, format }: ArgumentElement) {
-  let argumentType = getArgumentType(format);
-  let offset = (format as PluralFormat).offset
-    ? `, offset:${(format as PluralFormat).offset}`
-    : '';
+function printSimpleFormatElement(
+  el: DateElement | TimeElement | NumberElement
+) {
+  return `{${el.value}, ${el.type}${el.style ? `, ${el.style}` : ''}}`;
+}
+function printSelectElement(el: SelectElement) {
+  const msg = [
+    el.value,
+    'select',
+    Object.keys(el.options)
+      .map(id => `${id}{${_printICUMessage(el.options[id].value)}}`)
+      .join(' ')
+  ].join(',');
+  return `{${msg}}`;
+}
 
-  let options = (format as PluralFormat).options.map(option => {
-    let optionValue = _printICUMessage(option.value);
-    return ` ${option.selector} {${optionValue}}`;
-  });
-
-  return `{${id}, ${argumentType}${offset},${options.join('')}}`;
+function printPluralElement(el: PluralElement) {
+  const type = el.pluralType === 'cardinal' ? 'plural' : 'selectordinal';
+  const msg = [
+    el.value,
+    type,
+    el.offset ? `offset:${el.offset}` : '',
+    Object.keys(el.options)
+      .map(id => `${id}{${_printICUMessage(el.options[id].value)}}`)
+      .join(' ')
+  ].join(',');
+  return `{${msg}}`;
 }