--- a/packages/intl-messageformat-parser/.gitignore
+++ b/packages/intl-messageformat-parser/.gitignore
@@ -1,2 +1,2 @@
 lib/
-src/parser.js
+src/parser.ts