--- a/packages/intl-messageformat-parser/.gitignore
+++ b/packages/intl-messageformat-parser/.gitignore
@@ -1,2 +1,3 @@
 lib/
 src/parser.ts
+src/parser.js