--- a/packages/intl-messageformat-parser/build.js
+++ b/packages/intl-messageformat-parser/build.js
@@ -1,14 +1,54 @@
 #!/usr/bin/env node
 const peg = require('pegjs')
+const tspegjs = require("ts-pegjs");
 const fs = require('fs')
 const {outputFileSync} = require('fs-extra')
 const grammar = fs.readFileSync('./src/parser.pegjs', 'utf-8')
 
-// ES6
-outputFileSync('src/parser.js', `export default ${peg.generate(grammar, {output: 'source'})}`)
+// TS
+const srcString = peg.generate(grammar, {
+    plugins: [tspegjs],
+    output: 'source',
+    tspegjs: {
+        customHeader: `
+import { 
+    MessageFormatElement, 
+    LiteralElement, 
+    ArgumentElement, 
+    NumberElement, 
+    DateElement, 
+    TimeElement, 
+    SelectElement, 
+    PluralElement,
+    TYPE
+} from './types'`
+    },
+    returnTypes: {
+        argument: 'string',
+        ws: 'string',
+        digit: 'string',
+        hexDigit: 'string',
+        quoteEscapedChar: 'string',
+        apostrophe: 'string',
+        escape: 'string',
+        char: 'string',
+        chars: 'string',
+        varName: 'string',
+        number: 'number',
+        start: 'MessageFormatElement[]',
+        message: 'MessageFormatElement[]',
+        literalElement: 'LiteralElement',
+        argumentElement: 'ArgumentElement',
+        selectElement: 'SelectElement',
+        pluralElement: 'PluralElement',
+        simpleFormatElement: `
+| NumberElement 
+| DateElement 
+| TimeElement 
+`
+    }
+})
 
-// Globals
-outputFileSync('dist/parser.js', peg.generate(grammar, {output: 'source', format: 'globals', exportVar: 'IntlMessageFormatParser'}))
+const REGEX = /ParseFunction = \((.*?)\) => (any);/g
 
-// CJS
-outputFileSync('lib/parser.js', peg.generate(grammar, {output: 'source', format: 'commonjs'}))
+outputFileSync('src/parser.ts', srcString.replace(REGEX, 'ParseFunction = ($1) => MessageFormatElement[];'))