--- a/packages/intl-messageformat-parser/index.js
+++ b/packages/intl-messageformat-parser/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var parser = require('./lib/parser')
+var parser = require('./dist/parser')
 
 module.exports = parser
 module.exports['default'] = parser