--- a/packages/intl-messageformat-parser/jest.config.js
+++ b/packages/intl-messageformat-parser/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tsconfig.cjs.json'
+    }
+  }
+};