--- a/packages/intl-messageformat-parser/src/index.ts
+++ b/packages/intl-messageformat-parser/src/index.ts
@@ -1,15 +1,18 @@
-import {parse as pegParse, IParseOptions} from './parser'
-import {Options, MessageFormatElement} from './types'
+import { parse as pegParse, IParseOptions } from './parser';
+import { Options, MessageFormatElement } from './types';
 import { normalizeHashtagInPlural } from './normalize';
 export * from './types';
 export * from './parser';
 
-export type ParseOptions = Options & IParseOptions
+export type ParseOptions = Options & IParseOptions;
 
-export function parse(input: string, opts?: ParseOptions): MessageFormatElement[] {
-    const els = pegParse(input, opts)
-    if (!opts || opts.normalizeHashtagInPlural !== false) {
-        normalizeHashtagInPlural(els)
-    }
-    return els
+export function parse(
+  input: string,
+  opts?: ParseOptions
+): MessageFormatElement[] {
+  const els = pegParse(input, opts);
+  if (!opts || opts.normalizeHashtagInPlural !== false) {
+    normalizeHashtagInPlural(els);
+  }
+  return els;
 }