--- a/packages/intl-messageformat-parser/src/index.ts
+++ b/packages/intl-messageformat-parser/src/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './parser';