--- a/packages/intl-messageformat-parser/src/parser.pegjs
+++ b/packages/intl-messageformat-parser/src/parser.pegjs
@@ -5,8 +5,8 @@ See the accompanying LICENSE file for terms.
 */
 
 /*
-`TYPE` comes from `types.ts`
-*/
+ * IMPORTANT: `TYPE` comes from `types.ts`
+ */
 
 {
     function insertLocation () {