--- a/packages/intl-messageformat/src/index.ts
+++ b/packages/intl-messageformat/src/index.ts
@@ -4,12 +4,10 @@ Copyrights licensed under the New BSD License.
 See the accompanying LICENSE file for terms.
 */
 
-import parser from 'intl-messageformat-parser';
+import {parse} from 'intl-messageformat-parser';
 import IntlMessageFormat from './core';
 
-IntlMessageFormat.__parse = parser.parse;
+IntlMessageFormat.__parse = parse;
 
-export { Formats, Pattern } from './compiler';
 export * from './core';
-export { Formatters } from './compiler';
 export default IntlMessageFormat;