--- a/packages/intl-messageformat/src/index.ts
+++ b/packages/intl-messageformat/src/index.ts
@@ -4,7 +4,7 @@ Copyrights licensed under the New BSD License.
 See the accompanying LICENSE file for terms.
 */
 
-import {parse} from 'intl-messageformat-parser';
+import { parse } from 'intl-messageformat-parser';
 import IntlMessageFormat from './core';
 
 IntlMessageFormat.__parse = parse;