BREAKING CHANGE: This completely changes the AST produced by the parser

Before:
```
complex_msg AST length 12567
normal_msg AST length 2638
simple_msg AST length 567
string_msg AST length 288
complex_msg x 3,405 ops/sec ±5.44% (81 runs sampled)
normal_msg x 27,513 ops/sec ±2.14% (87 runs sampled)
simple_msg x 113,043 ops/sec ±1.20% (89 runs sampled)
string_msg x 147,838 ops/sec ±0.78% (90 runs sampled)
```

After:
```
complex_msg AST length 2053
normal_msg AST length 410
simple_msg AST length 79
string_msg AST length 36
complex_msg x 3,926 ops/sec ±2.37% (90 runs sampled)
normal_msg x 27,641 ops/sec ±3.93% (86 runs sampled)
simple_msg x 100,764 ops/sec ±5.35% (79 runs sampled)
string_msg x 120,362 ops/sec ±7.11% (74 runs sampled)
```

Also speeds up cold case for `intl-messageformat`:
Before
```
intl-messageformat: > intl-messageformat@4.4.0 benchmark /Users/longho/src/formatjs/packages/intl-messageformat
intl-messageformat: > ts-node --project ./tsconfig.cjs.json tests/benchmark
intl-messageformat: format_cached_complex_msg x 464,349 ops/sec ±1.55% (92 runs sampled)
intl-messageformat: format_cached_string_msg x 130,717,683 ops/sec ±8.70% (89 runs sampled)
intl-messageformat: new_complex_msg_preparsed x 2,994 ops/sec ±17.02% (77 runs sampled)
intl-messageformat: new_complex_msg x 1,388 ops/sec ±15.39% (72 runs sampled)
intl-messageformat: new_string_msg x 44,845 ops/sec ±4.11% (83 runs sampled)
intl-messageformat: complex msg format x 1,594 ops/sec ±9.89% (81 runs sampled)
intl-messageformat: complex msg w/ formatters format x 2,017 ops/sec ±16.10% (73 runs sampled)
intl-messageformat: complex preparsed msg w/ formatters format x 31,138 ops/sec ±4.14% (84 runs sampled)
intl-messageformat: complex preparsed msg w/ new formatters format x 33,853 ops/sec ±1.71% (82 runs sampled)
```

After
```
intl-messageformat: > intl-messageformat@4.4.0 benchmark /Users/longho/src/formatjs/packages/intl-messageformat
intl-messageformat: > ts-node --project ./tsconfig.cjs.json tests/benchmark
intl-messageformat: format_cached_complex_msg x 216,154 ops/sec ±2.01% (90 runs sampled)
intl-messageformat: format_cached_string_msg x 140,835,947 ops/sec ±0.87% (93 runs sampled)
intl-messageformat: new_complex_msg_preparsed x 8,295 ops/sec ±12.17% (78 runs sampled)
intl-messageformat: new_complex_msg x 2,386 ops/sec ±16.66% (83 runs sampled)
intl-messageformat: new_string_msg x 49,049 ops/sec ±2.00% (91 runs sampled)
intl-messageformat: complex msg format x 2,629 ops/sec ±1.80% (86 runs sampled)
intl-messageformat: complex msg w/ formatters format x 3,543 ops/sec ±2.25% (85 runs sampled)
intl-messageformat: complex preparsed msg w/ formatters format x 49,798 ops/sec ±5.09% (89 runs sampled)
intl-messageformat: complex preparsed msg w/ new formatters format x 51,529 ops/sec ±1.14% (89 runs sampled)
```